refactor: Move custom exception classes to domain layer (#129)
This commit relocates custom exception classes from the `common.exception` package to their respective domain-specific packages within the `domain` layer.
