--- app/src/main/java/com/example/rentit/common/exception/product/NotProductOwnerException.kt
@@ -1,3 +1,0 @@
-package com.example.rentit.common.exception.product
-
-class NotProductOwnerException: Exception()
