--- app/src/main/java/com/example/rentit/common/exception/product/ResvByOwnerException.kt
@@ -1,3 +1,0 @@
-package com.example.rentit.common.exception.product
-
-class ResvByOwnerException(): Exception()
