--- app/src/main/java/com/example/rentit/common/exception/rental/RentalNotFoundException.kt
@@ -1,3 +1,0 @@
-package com.example.rentit.common.exception.rental
-
-class RentalNotFoundException: Exception()
