+++ app/src/main/java/com/example/rentit/domain/chat/exception/ChatRoomAlreadyExistsException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.common.exception.chat
+package com.example.rentit.domain.chat.exception
 
 class ChatRoomAlreadyExistsException: Exception()
