+++ app/src/main/java/com/example/rentit/domain/chat/exception/ForbiddenChatAccessException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.common.exception.chat
+package com.example.rentit.domain.chat.exception
 
 class ForbiddenChatAccessException(message: String? = null): Exception(message)
