+++ app/src/main/java/com/example/rentit/domain/product/exception/NotProductOwnerException.kt
@@ -1,0 +1,3 @@
+package com.example.rentit.domain.product.exception
+
+class NotProductOwnerException: Exception()
