+++ app/src/main/java/com/example/rentit/domain/product/exception/ResvByOwnerException.kt
@@ -1,0 +1,3 @@
+package com.example.rentit.domain.product.exception
+
+class ResvByOwnerException(): Exception()
