+++ app/src/main/java/com/example/rentit/domain/rental/exception/EmptyBodyException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.common.exception.rental
+package com.example.rentit.domain.rental.exception
 
 class EmptyBodyException(message: String = "Empty response body"): Exception(message)
