+++ app/src/main/java/com/example/rentit/domain/rental/exception/RentalNotFoundException.kt
@@ -1,0 +1,3 @@
+package com.example.rentit.domain.rental.exception
+
+class RentalNotFoundException: Exception()
