+++ app/src/main/java/com/example/rentit/domain/rental/exception/TooManyRequestException.kt
@@ -1,0 +1,3 @@
+package com.example.rentit.domain.rental.exception
+
+class TooManyRequestException: Exception()
