+++ app/src/main/java/com/example/rentit/domain/rental/exception/VerificationRequestNotFoundException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.common.exception.rental
+package com.example.rentit.domain.rental.exception
 
 class VerificationRequestNotFoundException : Exception()
