+++ app/src/main/java/com/example/rentit/domain/user/exception/GoogleSsoFailedException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.common.exception.user
+package com.example.rentit.domain.user.exception
 
 class GoogleSsoFailedException(message: String? = null): Exception(message)
