+++ app/src/main/java/com/example/rentit/presentation/auth/login/LoginViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.rentit.BuildConfig
 import com.example.rentit.common.exception.MissingTokenException
-import com.example.rentit.common.exception.user.GoogleSsoFailedException
+import com.example.rentit.domain.user.exception.GoogleSsoFailedException
 import com.example.rentit.domain.user.repository.UserRepository
 import com.example.rentit.domain.user.usecase.InitializeAuthUseCase
 import com.google.android.gms.auth.api.signin.GoogleSignIn
